+++ Nos/Models/CoreData/Relay+CoreDataClass.swift
@@ -12,11 +12,10 @@ public class Relay: NosManagedObject {
 
     static var recommended: [String] {
         [
+        nosAddress.absoluteString,
         "wss://relay.damus.io",
         "wss://purplepag.es",
-        nosAddress.absoluteString,
-        "wss://relay.snort.social",
-        "wss://olympics2024.nos.social",
+        "wss://nos.lol",
         ]
     }